login server update

+++ applications/common/include/common/net/packet/login_togateway/playerleave.h
@@ -49,7 +49,7 @@ class PlayerLeave : public pf_net::packet::Base {
  private:
    char account_[ACCOUNT_LENGTH_MAX];
    int16_t centerid_;
-   int16_t guid_;
+   int64_t guid_;
 
 };
